Fix CodeSegment probing in sandboxed processes (#290)
Like com.apple.WebKit.* processes on iOS 11.x.

By allowing gum_file_open_tmp() to fail, we let try_realize() fail and
take the alternative path before giving up.

+++ gum/backend-darwin/gumcodesegment-darwin.c
@@ -273,6 +273,8 @@ gum_code_segment_try_realize (GumCodeSegment * self)
   fsignatures_t sigs;
 
   self->fd = gum_file_open_tmp ("frida-XXXXXX.dylib", &dylib_path);
+  if (self->fd == -1)
+    return FALSE;
 
   unlink (dylib_path);
 
@@ -572,7 +574,6 @@ gum_file_open_tmp (const gchar * tmpl,
     g_free (path);
     path = g_build_filename ("/Library/Caches", tmpl, NULL);
     res = mkstemps (path, suffix_length);
-    g_assert (res != -1);
   }
 
   *name_used = path;
